requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
DateTime==5.4
fonttools==4.46.0
googlesearch-python==1.2.3
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.8
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.1.0
zope.interface==6.1
